# docs/swagger.yaml
definitions:
...
  customerentity.Customer:
    properties:
      address:
        type: string
      customer_id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  entity.Inventory:
    properties:
      created_at:
        type: string
      id:
        type: integer
      product_id:
        type: integer
      stock:
        type: integer
      updated_at:
        type: string
      warehouse_id:
        type: integer
    type: object
  entity.InventoryStockOnly:
    properties:
      stock:
        type: integer
    type: object
...
  entity.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /customers:
    post:
      consumes:
      - application/json
      description: Add an Customer to the database
      parameters:
      - description: Customer data
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/customerentity.Customer'
      produces:
      - application/json
      responses:
        "201":
          description: Customer created
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid customer_id format, Unable to parse form data, Unable
            to get Customer from form
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application AddCustomer error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Add customer
      tags:
      - Customer
  /customers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a customer from the database by ID
      parameters:
      - description: customer_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Customer deleted
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid customer_id DeleteCustomer
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application DeleteCustomer error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Delete customer
      tags:
      - Customer
    get:
      consumes:
      - application/json
      description: Get Customer details by customer_id
      parameters:
      - description: customer_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get customers
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid customer_id GetCustomer
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application GetCustomer error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Get customers
      tags:
      - Customer
    put:
      consumes:
      - application/json
      description: Update a Customer in the database by ID
      parameters:
      - description: customer_id
        in: path
        name: id
        required: true
        type: integer
      - description: Customer data
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/customerentity.Customer'
      produces:
      - application/json
      responses:
        "201":
          description: Customer updated
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid customer_id
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application UpdateCustomer error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Update customer
      tags:
      - Customer
  /images:
    post:
      consumes:
      - multipart/form-data
      description: Add an image to the database
      parameters:
      - description: Product ID
        in: formData
        name: product_id
        required: true
        type: integer
      - description: Caption
        in: formData
        name: caption
        type: string
      - description: Image file
        in: formData
        name: image_file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Image created
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid product ID format, Unable to parse form data, Unable
            to get image from form
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application AddImage error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Add image
      tags:
      - Image
  /images/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an image from the database by ID
      parameters:
      - description: image_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Image deleted
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid image ID DeleteImage
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application DeleteImage error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Delete image
      tags:
      - Image
    get:
      consumes:
      - application/json
      description: Get image details by image_id
      parameters:
      - description: image_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get images
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid image_id GetImage
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application GetImage error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Get image
      tags:
      - Image
  /inventories/{id}:
    get:
      consumes:
      - application/json
      description: Get inventory details by product ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Inventory details
          schema:
            $ref: '#/definitions/entity.Inventory'
        "400":
          description: Invalid product ID GetInventory
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application GetInventory error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Get inventory
      tags:
      - Inventory
    put:
      consumes:
      - application/json
      description: Update stock details for a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Stock details to update
        in: body
        name: Object
        required: true
        schema:
          $ref: '#/definitions/entity.InventoryStockOnly'
      produces:
      - application/json
      responses:
        "201":
          description: Stock updated
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid product ID UpdateStock or Invalid JSON
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application UpdateStock error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Update stock
      tags:
      - Inventory
...
swagger: "2.0"